move combiners into separate package
