+++ graphql-code-generator/src/main/java/io/ballerina/graphql/generator/service/generator/ServiceCodeGenerator.java
@@ -28,8 +28,8 @@ import io.ballerina.graphql.generator.CodeGeneratorConstants;
 import io.ballerina.graphql.generator.GenerationException;
 import io.ballerina.graphql.generator.GraphqlProject;
 import io.ballerina.graphql.generator.service.Constants;
-import io.ballerina.graphql.generator.service.ServiceCombiner;
-import io.ballerina.graphql.generator.service.ServiceFileCombiner;
+import io.ballerina.graphql.generator.service.combiner.ServiceCombiner;
+import io.ballerina.graphql.generator.service.combiner.ServiceFileCombiner;
 import io.ballerina.graphql.generator.service.exception.ServiceFileCombinerException;
 import io.ballerina.graphql.generator.service.exception.ServiceGenerationException;
 import io.ballerina.graphql.generator.service.exception.ServiceTypesFileCombinerException;
